Migrate functionality from CelestialSphereManager to MainUIController
