--- Assets/Scripts/CelestialSphereManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bbd7634d02e5435449fe15270310d222
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
